Modified PhysicalObject constructor to longer create Sprite based on SpritePath and Sprite dimensions, instead now takes a pointer to a Sprite in the parameter. Adjust the Constructor calls for all the derived classes.
Created Vectors (derived sets) of visualObject and physicalObject for efficient reference by Camera and Physics respectively.

Added new Abstract Class Entity and added as a Base Class to Player and EntityY.

Added TerrainObject in preparation for the Physics update.

Appended code to populate a Vector of TerrainObjects in conjunction with the existing TerrainMap.

Added new Function in Game to load (initiate) the Dervied Sets.

Reorganized PhysicalObject adding objectIsMovable attribute and adjusted the Gravity value.
(Also increased the jump value in Game)